Save spotify credentials instead of keeping the auth code
The auth code is a one-use only code, so it makes no sense to keep it in
the store. The whole process of retrieving the spotify credentials has
been integrated in one function and a new action has been added in order
to save such credentials in the store (e.g.: `access_token`,
`refresh_token`).