Keep track of the last active device
On every song change, the devices list will be updated. On every update
of the devices list, if there's an active device, it will be saved in
the redux state.

When trying to play a song, it might be that no device is currently
selected as active. In that case, the playback will be transferred to
the last active device.