Update playing status on play/pause
The `playing` key from the `currentTrack` object from the state is
updated instantly when triggering a `play` or `pause` action, instead of
waiting for a track update from the server which returns the current
state.

This way the action takes for granted that the play/pause request will
succeed (optimistic approach). In case it didn't work, the `playing`
status will be updated anyway next time the current track is updated (by
default every 5 seconds).

The advantage of doing this is that the play/pause button is always
updated instantly when clicking on it, improving thus the UX.